make reusable MuiThemeProvider
