+++ src/constants/index.js
@@ -5,7 +5,7 @@ export const publicVapidKey =
   'BOjfhJMEPQh3NggRpL_6ZTv4JWXsK8kHyo8xTJiUitepCzYDAg7d0cKuOiGclJgNrHKsFA7xQwuHJLj6icDp-5I';
 
 // images
-export const SPACE = `${process.env.PUBLIC_URL}/images/background.jpg`;
+export const GRADIENT = `${process.env.PUBLIC_URL}/images/background.jpg`;
 export const GOOGLE_LOGO = `${process.env.PUBLIC_URL}/images/google.png`;
 export const PICSART_LOGO = `${process.env.PUBLIC_URL}/images/picsart.jpg`;
 export const PICSART_LOGO_WHITE =
